Implement performance monitoring and profiling tools for Oxide CLI
- Added PerformanceMonitor struct for real-time performance analysis.
- Introduced PerformanceMetrics struct to track CPU, memory, disk I/O, and network usage.
- Created ProfilingSession and ProfileCheckpoint structs for detailed operation profiling.
- Implemented methods for starting and finishing profiling sessions, including checkpointing.
- Added functionality for continuous performance monitoring with configurable intervals.
- Developed benchmarking comparison logic to evaluate performance improvements or regressions.
- Included performance counter management for tracking custom metrics.
- Generated comprehensive performance reports and system health assessments.
- Added CLI commands for performance monitoring and profiling operations.
- Implemented unit tests for performance monitor creation, profiling sessions, significance calculation, and performance counters.

+++ ai/oxirs-shacl-ai/src/error_handling.rs
@@ -1342,7 +1342,7 @@ impl ErrorImpactAssessor {
             _ => 1.0,
         };
 
-        Ok((base_security_impact * criticality_multiplier).min(1.0))
+        Ok((base_security_impact as f64 * criticality_multiplier as f64).min(1.0_f64))
     }
 }
